move vendor code from './src/*' to './vendor/'
