+++ cmake/ETLSetupFeatures.cmake
@@ -133,31 +133,31 @@ if(BUILD_CLIENT)
 			"src/qcommon/i18n_main.cpp"
 			"src/qcommon/i18n_findlocale.c"
 			"src/qcommon/i18n_findlocale.h"
-			"src/tinygettext/tinygettext/dictionary.hpp"
-			"src/tinygettext/tinygettext/dictionary_manager.hpp"
-			"src/tinygettext/tinygettext/file_system.hpp"
-			"src/tinygettext/tinygettext/iconv.hpp"
-			"src/tinygettext/tinygettext/language.hpp"
-			"src/tinygettext/tinygettext/log.hpp"
-			"src/tinygettext/tinygettext/log_stream.hpp"
-			"src/tinygettext/tinygettext/plural_forms.hpp"
-			"src/tinygettext/tinygettext/po_parser.hpp"
-			"src/tinygettext/tinygettext/tinygettext.hpp"
-			"src/tinygettext/dictionary.cpp"
-			"src/tinygettext/dictionary_manager.cpp"
-			"src/tinygettext/iconv.cpp"
-			"src/tinygettext/language.cpp"
-			"src/tinygettext/log.cpp"
-			"src/tinygettext/plural_forms.cpp"
-			"src/tinygettext/po_parser.cpp"
-			"src/tinygettext/tinygettext.cpp"
+			"vendor/tinygettext/tinygettext/dictionary.hpp"
+			"vendor/tinygettext/tinygettext/dictionary_manager.hpp"
+			"vendor/tinygettext/tinygettext/file_system.hpp"
+			"vendor/tinygettext/tinygettext/iconv.hpp"
+			"vendor/tinygettext/tinygettext/language.hpp"
+			"vendor/tinygettext/tinygettext/log.hpp"
+			"vendor/tinygettext/tinygettext/log_stream.hpp"
+			"vendor/tinygettext/tinygettext/plural_forms.hpp"
+			"vendor/tinygettext/tinygettext/po_parser.hpp"
+			"vendor/tinygettext/tinygettext/tinygettext.hpp"
+			"vendor/tinygettext/dictionary.cpp"
+			"vendor/tinygettext/dictionary_manager.cpp"
+			"vendor/tinygettext/iconv.cpp"
+			"vendor/tinygettext/language.cpp"
+			"vendor/tinygettext/log.cpp"
+			"vendor/tinygettext/plural_forms.cpp"
+			"vendor/tinygettext/po_parser.cpp"
+			"vendor/tinygettext/tinygettext.cpp"
 		)
 		if(MSVC)
-			list(APPEND GETTEXT_SRC "${SRC}/tinygettext/windows_file_system.cpp")
-			list(APPEND GETTEXT_SRC "${SRC}/tinygettext/tinygettext/windows_file_system.hpp")
+			list(APPEND GETTEXT_SRC "${PROJECT_SOURCE_DIR}/vendor/tinygettext/windows_file_system.cpp")
+			list(APPEND GETTEXT_SRC "${PROJECT_SOURCE_DIR}/vendor/tinygettext/tinygettext/windows_file_system.hpp")
 		else()
-			list(APPEND GETTEXT_SRC "${SRC}/tinygettext/unix_file_system.cpp")
-			list(APPEND GETTEXT_SRC "${SRC}/tinygettext/tinygettext/unix_file_system.hpp")
+			list(APPEND GETTEXT_SRC "${PROJECT_SOURCE_DIR}/vendor/tinygettext/unix_file_system.cpp")
+			list(APPEND GETTEXT_SRC "${PROJECT_SOURCE_DIR}/vendor/tinygettext/tinygettext/unix_file_system.hpp")
 		endif()
 		target_sources(client_libraries INTERFACE ${GETTEXT_SRC})
 	endif(FEATURE_GETTEXT)
@@ -383,8 +383,8 @@ if(BUILD_MOD)
 	if(FEATURE_OMNIBOT)
 		target_compile_definitions(qagame_libraries INTERFACE FEATURE_OMNIBOT)
 		target_sources(qagame_libraries INTERFACE
-			"${SRC}/game/g_etbot_interface.cpp"
-			"${SRC}/Omnibot/Common/BotLoadLibrary.cpp"
+			"${PROJECT_SOURCE_DIR}/vendor/g_etbot_interface.cpp"
+			"${PROJECT_SOURCE_DIR}/vendor/Omnibot/Common/BotLoadLibrary.cpp"
 		)
 	endif()
