+++ src/qcommon/i18n_main.cpp
@@ -52,10 +52,10 @@ extern "C"
 #include <stdlib.h>
 #include <map>
 
-#include "../tinygettext/tinygettext/po_parser.hpp"
-#include "../tinygettext/tinygettext/tinygettext.hpp"
-#include "../tinygettext/tinygettext/log.hpp"
-#include "../tinygettext/tinygettext/file_system.hpp"
+#include "../../vendor/tinygettext/tinygettext/po_parser.hpp"
+#include "../../vendor/tinygettext/tinygettext/tinygettext.hpp"
+#include "../../vendor/tinygettext/tinygettext/log.hpp"
+#include "../../vendor/tinygettext/tinygettext/file_system.hpp"
 
 tinygettext::DictionaryManager dictionary;
 tinygettext::DictionaryManager dictionary_mod;
